% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniintervals.R
\name{bonf.T2}
\alias{bonf.T2}
\title{Compute Bonferroni-T2 Confidence Intervals}
\usage{
bonf.T2(x.mean, S, alpha, A, k = NULL)
}
\arguments{
\item{x.mean}{A numeric vector representing the sample means.}

\item{S}{A numeric matrix representing the sample covariance matrix.}

\item{alpha}{A numeric value representing the significance level for the confidence intervals.}

\item{A}{A numeric matrix where each row represents a linear combination of the means for which the confidence interval is computed.}

\item{k}{An optional integer representing the number of comparisons. If `NULL`, it defaults to the number of rows of `A`.}
}
\value{
A matrix with three columns:
\describe{
  \item{inf}{The lower bounds of the confidence intervals.}
  \item{center}{The point estimates for the linear combinations of the means.}
  \item{sup}{The upper bounds of the confidence intervals.}
}
}
\description{
This function computes Bonferroni-T2 confidence intervals for linear combinations of means.
}
\details{
The function calculates the Bonferroni-adjusted confidence intervals for the linear combinations of the means specified in the rows of `A`. The adjustment accounts for multiple comparisons to control the overall type I error rate.

If the provided `k` is greater than the number of rows in `A`, a warning is issued suggesting manual adjustments for other Bonferroni intervals. If `k` is smaller than the number of rows in `A`, a warning is issued and `k` is set to the number of rows in `A`.
}
\examples{
x.mean <- c(2.5, 3.0, 4.5)
S <- matrix(c(1, 0.5, 0.3,
              0.5, 2, 0.4,
              0.3, 0.4, 3), nrow = 3)
alpha <- 0.05
A <- matrix(c(1, 0, 0,
              0, 1, 0,
              0, 0, 1), nrow = 3)
bonf.T2(x.mean, S, alpha, A)

}
