% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minianova.R
\name{manova.bonferroni}
\alias{manova.bonferroni}
\title{Compute Bonferroni Intervals for MANOVA or ANOVA}
\usage{
manova.bonferroni(fit.manova, data, effects1, effects2 = NULL, k.bonf, alpha)
}
\arguments{
\item{fit.manova}{An object representing a fitted MANOVA or ANOVA model.}

\item{data}{A dataframe containing the response variable(s).}

\item{effects1}{A factor representing the first treatment effect.}

\item{effects2}{An optional factor representing the second treatment effect for two-way designs. Default is NULL.}

\item{k.bonf}{The Bonferroni correction factor. Must be provided.}

\item{alpha}{The significance level for the confidence intervals.}

\item{effect1.name}{A character string for the name of the first effect (currently unused in the function).}

\item{effects2.name}{A character string for the name of the second effect (currently unused in the function).}
}
\value{
A list containing the following elements:
  \item{tau1}{A dataframe of estimated effects for the first treatment.}
  \item{tau2}{A dataframe of estimated effects for the second treatment (if applicable).}
  \item{general.mean}{A vector of overall means for each response variable.}
  \item{p}{The number of response variables (1 for ANOVA, >1 for MANOVA).}
  \item{ci.diff.tau1}{A dataframe of confidence intervals for differences between effects of the first treatment.}
  \item{ci.diff.tau2}{A dataframe of confidence intervals for differences between effects of the second treatment (if applicable).}
  \item{ci.effect.tau1}{A dataframe of confidence intervals for effects of the first treatment.}
  \item{ci.effect.tau2}{A dataframe of confidence intervals for effects of the second treatment (if applicable).}
}
\description{
This function calculates Bonferroni intervals for Multivariate Analysis of Variance (MANOVA) 
or univariate Analysis of Variance (ANOVA). It can handle one-way or two-way designs.
The function provides confidence intervals for differences between treatment effects 
and for the effects themselves.
}
\details{
This function can be used for both MANOVA and ANOVA:
- For ANOVA, the input data should have a single response variable (p = 1).
- For MANOVA, the input data should have multiple response variables (p > 1).

The function automatically adapts to the number of response variables in the input data.

For one-way designs, use only the `effects1` parameter. For two-way designs, 
use both `effects1` and `effects2`.
}
\examples{
\dontrun{
# For one-way ANOVA
anova_results <- manova.bonferroni(fit.anova, my_data, effects1 = factor(treatment),
                                   k.bonf = 3, alpha = 0.05)

# For one-way MANOVA
manova_results <- manova.bonferroni(fit.manova, my_data, effects1 = factor(treatment),
                                    k.bonf = 4, alpha = 0.05)

# For two-way ANOVA
two_way_anova_results <- manova.bonferroni(fit.anova, my_data, 
                                           effects1 = factor(treatment1),
                                           effects2 = factor(treatment2),
                                           k.bonf = 6, alpha = 0.05)

# For two-way MANOVA
two_way_manova_results <- manova.bonferroni(fit.manova, my_data, 
                                            effects1 = factor(treatment1),
                                            effects2 = factor(treatment2),
                                            k.bonf = 8, alpha = 0.05)
                                            }

}
