% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiniR-package.R
\docType{package}
\name{MiniR-package}
\alias{MiniR}
\alias{MiniR-package}
\title{MiniR: Utility functions for Applied Statistics Exam}
\description{
A more detailed description of what your package does.
}
\keyword{internal}
