% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniplot.R
\name{pc.interpretation}
\alias{pc.interpretation}
\title{Interpret Principal Component Analysis Results}
\usage{
pc.interpretation(pc.result, data, number_pcs = 3)
}
\arguments{
\item{pc.result}{A list object containing the PCA result, typically obtained from the \code{\link{prcomp}} or \code{\link{princomp}} functions. It should include at least \code{$sdev} and \code{$loadings} components.}

\item{data}{A data frame or matrix containing the original data that was used for the PCA. This data is used to compute the variances of the original variables.}

\item{number_pcs}{An integer indicating the number of principal components to plot the loadings for. The default is 3.}
}
\value{
This function does not return any value. It generates plots as a side effect.
}
\description{
This function generates a comprehensive set of plots to aid in the interpretation of principal component analysis (PCA) results. It produces three main plots: the variance of the principal components, the variance of the original variables, and the cumulative contribution of the principal components to the total variance. Additionally, it provides barplots of the loadings for a specified number of principal components.
}
\details{
The function generates the following plots:
\itemize{
  \item A scree plot showing the variances of the principal components.
  \item A bar plot showing the variances of the original variables.
  \item A line plot showing the cumulative contribution of the principal components to the total variance.
  \item Bar plots showing the loadings for each of the specified principal components.
}
}
\note{
The function modifies the graphical parameters using \code{par} to arrange multiple plots on one page. It restores the original parameters after plotting.
}
\examples{
\dontrun{
data(iris)
pca_result <- prcomp(iris[, -5], scale. = TRUE)
pc.interpretation(pca_result, iris[, -5])
}

}
