% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniplot.R
\name{plot.intervals}
\alias{plot.intervals}
\title{Plot Confidence Intervals}
\usage{
\method{plot}{intervals}(BonfCI = NULL, SimCI = NULL, delta.0, legend.position = "bottomright")
}
\arguments{
\item{BonfCI}{A matrix or vector representing the Bonferroni confidence intervals. If a vector is provided, it will be reshaped into a matrix with 3 columns.}

\item{SimCI}{A matrix or vector representing the Simultaneous T2 confidence intervals. If a vector is provided, it will be reshaped into a matrix with 3 columns.}

\item{delta.0}{A vector representing the hypothesized values under H0.}

\item{legend.position}{A character string indicating the position of the legend. Default is 'bottomright'.}
}
\value{
A plot showing the confidence intervals.
}
\description{
This function plots Bonferroni and/or Simultaneous T2 confidence intervals for a given set of data.
}
\details{
The function plots the confidence intervals provided. If both `BonfCI` and `SimCI` are given, both sets of intervals will be plotted. If only one is provided, only that set of intervals will be plotted.

- The Bonferroni intervals are plotted in orange.
- The Simultaneous T2 intervals are plotted in blue.
- The hypothesized values under H0 are plotted in black.
}
\examples{
# Example usage with both Bonferroni and Simultaneous T2 intervals
BonfCI <- matrix(c(1, 2, 3, 2, 3, 4, 1, 3, 4), ncol = 3, byrow = TRUE)
SimCI <- matrix(c(1.5, 2.5, 3.5, 2.5, 3.5, 4.5, 1.5, 3.5, 4.5), ncol = 3, byrow = TRUE)
delta.0 <- c(2, 3, 3)
plot.intervals(BonfCI, SimCI, delta.0)

}
