% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniplot.R
\name{plot_decision_boundaries}
\alias{plot_decision_boundaries}
\title{Plot Decision Boundaries for LDA and QDA Models}
\usage{
plot_decision_boundaries(formula, data, fit.result)
}
\arguments{
\item{formula}{A formula specifying the model. The left-hand side of the formula should be the response variable, and the right-hand side should be the predictor variables.}

\item{data}{A data frame containing the variables in the formula.}

\item{fit.result}{A fitted model object of class \code{lda} or \code{qda} from the \code{\link[MASS]{lda}} or \code{\link[MASS]{qda}} functions, respectively.}
}
\value{
A ggplot object displaying the decision boundaries and the data points. Points are colored by their true class, and the background is shaded according to the predicted class.
}
\description{
This function plots the decision boundaries for Linear Discriminant Analysis (LDA) and Quadratic Discriminant Analysis (QDA) models using ggplot2. It visualizes the separation of different classes in the feature space based on the model's predictions.
}
\details{
The function works as follows:
\itemize{
  \item It extracts the feature variables and response variable from the formula and data.
  \item It creates a grid of points spanning the feature space.
  \item It predicts the class for each point in the grid using the provided LDA or QDA model.
  \item It uses ggplot2 to plot the decision boundaries and data points.
}
}
\note{
The function currently supports only two feature variables.
}
\examples{
\dontrun{
library(MASS)
library(ggplot2)
library(gridExtra)

data(iris)
lda_model <- lda(Species ~ Sepal.Length + Sepal.Width, data = iris)
plot_decision_boundaries(Species ~ Sepal.Length + Sepal.Width, iris, lda_model)
}

}
