% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniintervals.R
\name{sim.T2}
\alias{sim.T2}
\title{Simultaneous Confidence Intervals based on Hotelling's T-squared distribution}
\usage{
sim.T2(x.mean, S, alpha, A)
}
\arguments{
\item{x.mean}{A numeric vector representing the sample means.}

\item{S}{A numeric matrix representing the sample covariance matrix.}

\item{alpha}{A numeric value representing the significance level (e.g., 0.05 for 95% confidence).}

\item{A}{A numeric matrix where each row represents a linear combination of the means.}
}
\value{
A numeric matrix with three columns: `inf` for the lower bound of the SCI, `center` for the 
point estimates of the linear combinations, and `sup` for the upper bound of the SCI. Each row corresponds 
to a linear combination defined by a row of the matrix `A`.
}
\description{
This function calculates the simultaneous confidence intervals (SCI) for linear combinations of means 
based on Hotelling's T-squared distribution.
}
\details{
This function computes the simultaneous confidence intervals using Hotelling's T-squared distribution. 
The critical value is obtained from the Fisher's F distribution. The simultaneous confidence intervals are 
computed for the linear combinations specified by the rows of the matrix `A`.
}
\examples{
x.mean <- c(5, 10)
S <- matrix(c(4, 2, 2, 3), nrow = 2)
alpha <- 0.05
A <- matrix(c(1, 0, 0, 1), nrow = 2)
sim.T2(x.mean, S, alpha, A)

}
