% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniplot.R
\name{plot_decision_boundaries}
\alias{plot_decision_boundaries}
\title{Plot Decision Boundaries for LDA or QDA Models}
\usage{
plot_decision_boundaries(formula, data, fit.result)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables in the formula.}

\item{fit.result}{An object of class \code{lda} or \code{qda}, representing the fitted model.}
}
\value{
A ggplot object showing the decision boundaries and the data points.
}
\description{
This function creates a plot of decision boundaries for Linear Discriminant Analysis (LDA) or Quadratic Discriminant Analysis (QDA) models using ggplot2.
}
\examples{
\dontrun{
library(MASS)
library(ggplot2)
library(gridExtra)
data(iris)
fit <- lda(Species ~ Sepal.Length + Sepal.Width, data = iris)
plot_decision_boundaries(Species ~ Sepal.Length + Sepal.Width, iris, fit)
}

}
